Move ExecutionMode to own file
