+++ osu.Framework/Platform/ExecutionMode.cs
@@ -1,0 +1,16 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.ComponentModel;
+
+namespace osu.Framework.Platform
+{
+    public enum ExecutionMode
+    {
+        [Description("Single thread")]
+        SingleThread,
+
+        [Description("Multithreaded")]
+        MultiThreaded
+    }
+}
